requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
drf-yasg
django-filter
django-cors-middleware
gunicorn
pillow